+++ packages/nano/dio.js
@@ -34,12 +34,20 @@
 	
 	
 	// current version
-	var version = '6.0.1';
+	var version = '6.0.2';
 	
-	// enviroment specific variables
+	// enviroment
 	var document = window.document || null;
 	var browser = document !== null;
+	
+	// other
 	var Promise = window.Promise;
+	var requestAnimationFrame = window.requestAnimationFrame;
+	var setImmediate = window.setImmediate;
+	
+	var promiseSupport = Promise !== void 0;
+	var requestAnimationFrameSupport = requestAnimationFrame !== void 0;
+	var setImmediateSupport = setImmediate !== void 0;
 	
 	// namespaces
 	var nsStyle = 'data-scope';
@@ -50,21 +58,38 @@
 	// empty shapes
 	var objEmpty = {};
 	var arrEmpty = [];
-	var nodeEmpty = createVNodeShape(0, '', objEmpty, arrEmpty, null, null, null, null, null);
+	var nodeEmpty = createNodeShape(0, '', objEmpty, arrEmpty, null, null, null, null, null);
 	var funcEmpty = function () {};
 	var fragProps = {style: 'display: inherit;'};
 	
 	
 	/**
-	 * call function with context
+	 * schedule callback
 	 * 
-	 * @param  {function} func
-	 * @param  {any}      context
-	 * @param  {any}      arg
-	 * @return {any}
+	 * @type {function}
+	 * @param {function} callback
 	 */
-	function call (func, context, arg) {
-		return func.call(context, arg);
+	var schedule;
+	
+	if (requestAnimationFrameSupport) {
+		schedule = function schedule (callback) { 
+			requestAnimationFrame(callback);
+		}
+	}
+	else if (setImmediateSupport) {
+		schedule = function schedule (callback) { 
+			setImmediate(callback); 
+		}
+	}
+	else if (promiseSupport) {
+		schedule = function schedule (callback) { 
+			Promise.resolve().then(callback); 
+		}
+	}
+	else {
+		schedule = function schedule (callback) { 
+			setTimeout(callback, 0); 
+		}
 	}
 	
 	
@@ -190,7 +215,7 @@
 	
 	
 	/**
-	 * create VNode shape
+	 * create node shape
 	 *
 	 * @param  {number}                      nodeType
 	 * @param  {(string|function|Component)} type
@@ -202,7 +227,7 @@
 	 * @param  {Component}                   parent
 	 * @return {VNode}
 	 */
-	function createVNodeShape (nodeType, type, props, children, DOMNode, instance, index, parent, key) {
+	function createNodeShape (nodeType, type, props, children, DOMNode, instance, index, parent, key) {
 		return {
 			nodeType: nodeType,
 			type: type,
@@ -237,7 +262,6 @@
 	}
 	
 	
-
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
@@ -1460,7 +1484,7 @@
 	 * @return {VNode}
 	 */
 	function cloneNode (subject) {
-		return createVNodeShape(
+		return createNodeShape(
 			subject.nodeType,
 			subject.type,
 			subject.props,
@@ -1609,7 +1633,7 @@
 			} 
 			// element root node
 			else {
-				patchNodes(newNode, oldNode, newType, oldType);
+				reconcileNodes(newNode, oldNode, newType, oldType);
 			}
 		}
 	
@@ -1925,7 +1949,9 @@
 	
 		// intial throw from render, retry once
 		if (thrown === 0 && browser && location === 'render') {
-			setTimeout(call, 0, component.forceUpdate, component, null);
+			schedule(function () {
+				component.forceUpdate(null);
+			});
 		}
 		// multiple render throws / non-render location
 		else {
@@ -1964,7 +1990,7 @@
 		    if (authored && location !== 'stylesheet') {	    	
 		    	// return render node
 		    	if (location === 'render' || location === 'element') {
-		    		if (typeof newNode.type === 'string') {
+		    		if (newNode != null && typeof newNode.type === 'string') {
 		    			if (/^[A-z]/g.exec(newNode.type) === null) {
 	    					console.error(
 	    						'Dio bailed out of rendering an error state.\n\n'+
@@ -1980,16 +2006,16 @@
 		    		return newNode;
 		    	}
 		    	// async replace render node
-		    	else if (browser && newNode != null && newNode !== true && newNode !== false) {	 
-					setTimeout(
-						replaceRootNode, 
-						0, 
-						extractVirtualNode(newNode), 
-						oldNode = component.vnode, 
-						newNode.nodeType, 
-						oldNode.nodeType, 
-						component
-					)
+		    	else if (browser && newNode != null && newNode !== true && newNode !== false) {
+		    		schedule(function () {
+		    			replaceRootNode(
+		    				extractVirtualNode(newNode), 
+		    				oldNode = component.vnode, 
+		    				newNode.nodeType, 
+		    				oldNode.nodeType, 
+		    				component
+	    				)
+		    		});
 		    	}
 		    }
 		}
@@ -1999,7 +2025,7 @@
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
-	 * render components
+	 * extract
 	 * 
 	 * ---------------------------------------------------------------------------------
 	 */
@@ -2369,7 +2395,7 @@
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
-	 * render props
+	 * props
 	 * 
 	 * ---------------------------------------------------------------------------------
 	 */
@@ -2385,27 +2411,13 @@
 	 */
 	function assignProps (target, props, onlyEvents, component) {
 		for (var name in props) {
-			assignProp(target, name, props, onlyEvents, component);
-		}
-	}
-	
-	
-	/**
-	 * assign prop for create element
-	 * 
-	 * @param  {Node}       target
-	 * @param  {string}     name
-	 * @param  {Object}     props
-	 * @param  {boolean}    onlyEvents,
-	 * @param  {Component}  component
-	 */
-	function assignProp (target, name, props, onlyEvents, component) {
-		if (isEventName(name)) {
-			addEventListener(target, extractEventName(name), props[name], component);
-		}
-		else if (onlyEvents === false) {
-			// add attribute
-			updateProp(target, 'setAttribute', name, props[name], props.xmlns);
+			if (isEventName(name)) {
+				addEventListener(target, extractEventName(name), props[name], component);
+			}
+			else if (onlyEvents === false) {
+				// add attribute
+				updateProp(target, 'setAttribute', name, props[name], props.xmlns);
+			}
 		}
 	}
 	
@@ -2417,82 +2429,41 @@
 	 * @param  {VNode} oldNode
 	 */
 	function patchProps (newNode, oldNode) {
-		var diff   = diffProps(newNode, oldNode, newNode.props.xmlns || '', []);
-		var length = diff.length;
+		var newProps = newNode.props;
+		var oldProps = oldNode.props;
+		var namespace = newNode.props.xmlns || '';
+		var target = oldNode.DOMNode;
+		var length = 0;
 	
-		// if diff length > 0 apply diff
-		if (length !== 0) {
-			var target = oldNode.DOMNode;
-	
-			for (var i = 0; i < length; i++) {
-				var prop = diff[i];
-				// [0: action, 1: name, 2: value, namespace]
-				updateProp(target, prop[0], prop[1], prop[2], prop[3]);
-			}
-	
-			oldNode.props = newNode.props;
-		}
-	}
-	
-	
-	/**
-	 * collect prop diffs
-	 * 
-	 * @param  {VNode}   newNode 
-	 * @param  {VNode}   oldNode 
-	 * @param  {string}  namespace
-	 * @param  {Array[]} diff
-	 * @return {Array[]}          
-	 */
-	function diffProps (newNode, oldNode, namespace, diff) {
 		// diff newProps
 		for (var newName in newNode.props) { 
-			diffNewProps(newNode, oldNode, newName, namespace, diff); 
+			var newValue = newProps[newName];
+			var oldValue = oldProps[newName];
+	
+			if (newValue != null && oldValue !== newValue) {
+				updateProp(target, 'setAttribute', newName, newValue, namespace);
+				
+				if (length === 0) {
+					length++;
+				}
+			}
 		}
 	
 		// diff oldProps
 		for (var oldName in oldNode.props) { 
-			diffOldProps(newNode, oldName, namespace, diff); 
+			var newValue = newProps[oldName];
+	
+			if (newValue == null) {
+				updateProp(target, 'removeAttribute', oldName, '', namespace);
+				
+				if (length === 0) {
+					length++;
+				}
+			}
 		}
 	
-		return diff;
-	}
-	
-	
-	/**
-	 * diff newProps agains oldProps
-	 * 
-	 * @param  {VNode}   newNode 
-	 * @param  {VNode}   oldNode 
-	 * @param  {string}  newName
-	 * @param  {string}  namespace
-	 * @param  {Array[]} diff
-	 * @return {Array[]}          
-	 */
-	function diffNewProps (newNode, oldNode, newName, namespace, diff) {
-		var newValue = newNode.props[newName];
-		var oldValue = oldNode.props[newName];
-	
-		if (newValue != null && oldValue !== newValue) {
-			diff[diff.length] = ['setAttribute', newName, newValue, namespace];
-		}
-	}
-	
-	
-	/**
-	 * diff oldProps agains newProps
-	 * 
-	 * @param  {VNode}   newNode 
-	 * @param  {Object}  oldName 
-	 * @param  {string}  namespace
-	 * @param  {Array[]} diff
-	 * @return {Array[]}          
-	 */
-	function diffOldProps (newNode, oldName, namespace, diff) {
-		var newValue = newNode.props[oldName];
-	
-		if (newValue == null) {
-			diff[diff.length] = ['removeAttribute', oldName, '', namespace];
+		if (length !== 0) {
+			oldNode.props = newNode.props;
 		}
 	}
 	
@@ -2546,7 +2517,12 @@
 		else {
 			// id, className etc..
 			if (targetProp !== void 0 && isSVG === false) {
-				target[propName] = propValue;
+				if (propName === 'style') {
+					target.style.cssText = propValue;
+				}
+				else {
+					target[propName] = propValue;
+				}
 			}
 			// setAttribute/removeAttribute
 			else {
@@ -2604,7 +2580,7 @@
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
-	 * render nodes
+	 * nodes
 	 * 
 	 * ---------------------------------------------------------------------------------
 	 */
@@ -2928,7 +2904,7 @@
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
-	 * render events
+	 * events
 	 * 
 	 * ---------------------------------------------------------------------------------
 	 */
@@ -3017,7 +2993,7 @@
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
-	 * render refs
+	 * refs
 	 * 
 	 * ---------------------------------------------------------------------------------
 	 */
@@ -3043,14 +3019,14 @@
 	/**
 	 * ---------------------------------------------------------------------------------
 	 * 
-	 * reconciler
+	 * reconcile
 	 * 
 	 * ---------------------------------------------------------------------------------
 	 */
 	
 	
 	/**
-	 * patch keyed nodes
+	 * reconcile keyed nodes
 	 *
 	 * @param {Object<string, any>[2]} keys
 	 * @param {Node}                   parentNode
@@ -3060,7 +3036,7 @@
 	 * @param {number}                 oldLength
 	 * @param {number}                 pos
 	 */
-	function patchKeys (keys, parentNode, newNode, oldNode, newLength, oldLength, pos) {
+	function reconcileKeys (keys, parentNode, newNode, oldNode, newLength, oldLength, pos) {
 		var reconciled = new Array(newLength);
 		var childNodes = parentNode.childNodes;
 	
@@ -3138,7 +3114,7 @@
 				} 
 				// append
 				else {
-					appendNode(nodeType, newChild, parentNode,nextNode);
+					appendNode(nodeType, newChild, parentNode, nextNode);
 				}
 	
 				added++;
@@ -3169,14 +3145,14 @@
 	}
 	
 	/**
-	 * patch nodes
+	 * reconcile nodes
 	 *  
 	 * @param  {VNode}  newNode
 	 * @param  {VNode}  oldNode
 	 * @param  {number} newNodeType
 	 * @param  {number} oldNodeType
 	 */
-	function patchNodes (newNode, oldNode, newNodeType, oldNodeType) {	
+	function reconcileNodes (newNode, oldNode, newNodeType, oldNodeType) {	
 		// if newNode and oldNode, exit early
 		if (newNode === oldNode) {
 			return;
@@ -3323,14 +3299,14 @@
 				}
 				// noop
 				else {
-					patchNodes(newChild, oldChild, newType, oldType);
+					reconcileNodes(newChild, oldChild, newType, oldType);
 				}
 			}
 	
 			// reconcile keyed children
 			if (keyed) {
 				// new and old keys object are of differing shapes
-				patchKeys([newKeys, oldKeys], parentNode, newNode, oldNode, newLength, oldLength, pos);
+				reconcileKeys([newKeys, oldKeys], parentNode, newNode, oldNode, newLength, oldLength, pos);
 			}
 		}