+++ src/Props/index.js
@@ -1,18 +1,14 @@
 /**
  * ---------------------------------------------------------------------------------
  * 
- * render props
+ * props
  * 
  * ---------------------------------------------------------------------------------
  */
 	
 
 import 'assignProps';
-import 'assignProp';
 import 'patchProps';
-import 'diffProps';
-import 'diffNewProps';
-import 'diffOldProps';
 import 'updateProp';
 import 'updatePropObject';
 import 'assignDefaultProps';