+++ src/Props/patchProps.js
@@ -5,19 +5,40 @@
  * @param  {VNode} oldNode
  */
 function patchProps (newNode, oldNode) {
-	var diff   = diffProps(newNode, oldNode, newNode.props.xmlns || '', []);
-	var length = diff.length;
+	var newProps = newNode.props;
+	var oldProps = oldNode.props;
+	var namespace = newNode.props.xmlns || '';
+	var target = oldNode.DOMNode;
+	var length = 0;
 
-	// if diff length > 0 apply diff
-	if (length !== 0) {
-		var target = oldNode.DOMNode;
+	// diff newProps
+	for (var newName in newNode.props) { 
+		var newValue = newProps[newName];
+		var oldValue = oldProps[newName];
 
-		for (var i = 0; i < length; i++) {
-			var prop = diff[i];
-			// [0: action, 1: name, 2: value, namespace]
-			updateProp(target, prop[0], prop[1], prop[2], prop[3]);
+		if (newValue != null && oldValue !== newValue) {
+			updateProp(target, 'setAttribute', newName, newValue, namespace);
+			
+			if (length === 0) {
+				length++;
+			}
 		}
+	}
 
+	// diff oldProps
+	for (var oldName in oldNode.props) { 
+		var newValue = newProps[oldName];
+
+		if (newValue == null) {
+			updateProp(target, 'removeAttribute', oldName, '', namespace);
+			
+			if (length === 0) {
+				length++;
+			}
+		}
+	}
+
+	if (length !== 0) {
 		oldNode.props = newNode.props;
 	}
 }