+++ src/Utilities/index.js
@@ -9,5 +9,5 @@
 
 import 'escape';
 import 'unicode';
-import 'call';
 import 'sandbox';
+import 'schedule';